Finished implementation of maze
The Maze 1 scene has been added. Contained within the scene is the fox with camera, the walls of the maze consisting of many rigid bodies, events which make use of triggers and moving walls, and a placeholder raven and giant mushroom sprite.

Created a new folder to hold all prefabs made for the maze. This contains CollisionTrigger, Moving wall, and TimeTrigger.

Moving wall is a rigid body wall which has the moving wall script attached to it allowing it to be moved by the various triggers

CollisionTrigger is a collision box with the CollisionTrigger script attached to it. This allows events to start based of the player colliding with it

TimeTrigger is a gameObject which just has the TimeTrigger script. This allows events to start after a certain amount of time has passed specified by the inspector.

Created a folder within scripts titled Scripts made for Maze, it contains CollisionTrigger, MazeTrigger, MovingWall, and TimeTrigger

MazeTrigger is an abstract class inherited by CollisionTrigger and TimeTrigger. It's purpose is to hold the TriggerEvent method which when called does a variety of things depending on what values an object inheriting in it holds in it's inspector. If the object has a wall, the wall will be moved to the position specified in the inspector. If the object has a player, then the player will be teleported in the x and y direction relative to it's current position. If the object has another trigger, it'll toggle that trigger's active status. If the object has another object it'll toggle that object's active status. If it's a oneshot then the trigger will disable itself after firing.

CollisionTrigger causes an event trigger as described in MazeTrigger whenever the object it is attached to collides with another object.

TimeTrigger causes an event trigger as described in MazeTrigger after the time set in the inspector passes. This might potentially have bugs when not set to trigger only once.

MovingWall provides the functionality to Moving walls so that they correctly move and stop at a specific place designated in a MazeTrigger's script.