Functional Hallucination Effect
Modifiable and functional hallucination effect, spores are made in prefabs to trigger it, an adjustable threshold to display the warning message, and character death when reaching maximum

The effect runs smoothly now, but I do found occasional cases where the engine still takes ages to render a frame if you don't actually run the game. But everything's good if you run the game.