Added Mushroom Animations and Reworked Spore
Added a new animations and animation controllers to the Animations Folder.

MiniMushroom is the animator controller for the mushroom enemy, allowing them to blink whenever they fire a spore.

miniMushroom is the animation clip for aforementioned behaviour

Spore is the animation controller for the spores allowing them to appears as if they're moving through the air

Mushroom Spores is the animation clip for the aforementioned

In the Sprites foldermoved Mini Mushroom Sprite Sheet to the SpriteSheets folder.

In the SpritesSheet folder

Added Spore Sprite Sheet to contain the sprites for the spore for use in animation, and

Mini Mushroom Spritesheet was altered to contain the sprites of both the open and closed eyed mushroom.

In the prefabs folder altered Mushroom Enemy Prefab and Spore Projectile

Mushroom Enemy prefab was changed to have an animator component using the Mini Mushroom Animator Controller.

Spore Projectile was changed to have an animator component with the Spore animation controller.

In the scripts folder, altered the Spore Spawner Script and the Spore Projectile Script

Spore Spawner Script was changed to have a few additional comments, and changed the timer system used to determine when the mushroom shot so that if it was on cooldown when a player left its radius, the cooldown would continue to countdown. The script was also slightly modified to allow for an animation to play when it shot.

SporeProjectileScript was significantly changed. Altered behaviour of spores to instead slowly home in on the player rather than fire and various speeds at the player. This was primarily done to ensure increase the threat of the mushrooms. Spores currently attempt to kill the player when touching them.