Added dialogue to Pig Enemy
Changed Pig Enemy prefab

Added sprite to pig enemy prefab so that it stops being invisible

Added the npc script to Pig Enemy

Added the dialogue package as a child of the Pig Enemy in the prefab so that the dialogue always worked no matter what scene the pig was placed in. Only thing to be aware of is that in the NPC script in the inspector the player needs to be set to the fox before the dialogue will work.

Dialogue works like any other so the player must be in contact with the pig.

Removed PIg dialogue scene, due to it no longer being necessary