Main and Farm fixed
Added a filter to main so that it should looks better now, features are added where going to the left edge of the scene will go to farm and right to the maze.
Sounds in main is fixed now where it plays different sources of footsteps depends on where the player is at(sand/grass), and a background sound. Weird white box is just UI for NPCs.
Farm is fixed now, it's just someone resets the speed of fox to 0 so that it won't move.